Add files via upload
In Bangladesh,SEC(Securities and Exchange Commission)
 rules have been distorted, which causes the consternation of sharehold
ers. There is no precise security system in BSEC(Bangladesh Securities
 and Exchange Commission). So there is a huge possibility of leaking out
 the formal data and information of BSEC as there is an improper se
curity system of their own. In these circumstances, it is necessary to
 maintain the system in a decentralized network security system to pro
tect it from different kinds of cyber attacks. At present, it is blockchain
 which provides numerous secured decentralized network which ensures
 the accountability and integrity of any system. By storing all SEC infor
mation in blockchain, no cyber-attack will be susceptible to stave it. This
 paper presents how to make secure all sensitive data of BSEC by using
 blockchain data storage while cryptographical functions will dominate
 all the data and information.